chore(demo): migrate to new material components
Mat components have been refactored in ng15. Their APIs were mostly kept
but their internal structure and css have changed. This allows for better
a11y and should adhere more to the material spec.

In my case, the usage of mat comps is not wide nor custom so the upgrade
was very easy.