feat: do not require media directive for custom views
Previously, when using custom viewer template, one had to attach media
directive to the img, video etc. so that the gallery registered whether
assets got loaded or errored. This is easy to miss and frustrating to
deal with, so I transformed the directive to an internal thing.

Now it unifies the loading by capturing the load/error events in capture
phase regardless of the assets being internal or external.