change: revamp sliding mechanism
This is one of the biggest changes that is bearly visible to the user.
Where to start?

I noticed a problem with previous & next items ever so slightly
overflowing into the current scrollport of the gallery viewer. It looked
bad so I investigated. It turned out pretty fast this was by design so
that next and prev images could be present, even if by a tiny bit, in
the scrollport, which would make them preloaded. I decided to remove
this and deal with the preloading later, but it didn't solve the problem
really.

It turns out, I still do math operations in JavaScript, where division
can turn into imprecise results and therefore unwanted overflows in the
gallery viewer. So I thought that, instead of setting the slider always
by raw pixels, I could use some help of CSS `calc()`.

This happens to be a great decisions in multiple ways. First of all, it
allows to perform calculations in more declarative, less manual way
compared to the previous JS only logic.

Secondly, and this is amazing, it allows me to use CSS var to depict
items' width without the need to query it via JS. This doesn't
completely eliminate the need, but one less usage is always great.

Thirdly, this has the unexpected benefit of not having to re-center
current item in the scrollport on resize or actually ever. Since the
width CSS var is dynamic, at least when relative like 100%, it always
adapts to the current scrollport no matter the width change, unlike raw
pixels! This is both a benefit for performance and for devs due to less
cognitive load.

And lastly, even though on the first glance the CSS transform of the
slider looks more complicated with the index, CSS var and tweak compared
to the raw pixels of the previous implementation, it is actually much
more readable and conveys the rationale of the current position much
better than a value like 2240px.

TODO
The lazy loading is currently broken due to the change, but I think the
time has come to start using the native loading="lazy", which is finally
supported everywhere!