feat: use native lazy loading instead of custom one
The custom loading worked more or less fine but it was only a temporary
solution before loading="lazy" got better browser support. That moment
has finally come and switching to it allows for removing a lot of code!

The downside is that write now I don't have a replacement for loading
the prev/next item. I think I could make it work by somehow exposing the
items on the sides in the scrollport, but I will deal with that once the
need arises.