chore: switch to nx
nx is far more sophisticated than ng-cli. The things I like about it is
the task caching, the separation of apps and libs (in the future I can
add other app eg for stackblitz) and integration with Cypress (not yet
present, but adding a component testing with Cy can be nice).