test(gallery): introduce cypress component testing
The plan is to use cypress for UI heavy testing, like checking the
position of the slider and such. I think this is the best option to test
as close to user's experience as possible.

The previous solution with Karma and headless Chrome just doesn't suffice
anymore - it is not easy to debug nor to simulate user's behavior. For
instance, to simulate user sliding, I have to manually dispatch events
on DOM nodes. Or I have to manually flush async tasks. That is not ideal
for maintenance, to put it mildly.

One note regarding webpack: For some reason, incorrect webpack types were
installed and there was a mismatch between expected and actual types.
Installing them explicitly solved the problem.