change!: turn item width into visible item count
This makes it declarative unlike before when the item count had to be
calculated from the item and viewer width, which not only saves code but
also resources.

One observation of mine is that this relatively simple change required
quite a few changes in the unit tests. It only shows that testing components
in Angular is pain (calling ngOnChanges, calling tick in fakeAsync etc.)
which makes a good argument for cy's component testing.