fix: prevent on init centering animation
This not only looked kinda bad since the images were moving with user's
interaction, but they also made the fringe elements visible long enough
for the to be loaded despite their loading="lazy".

This is not fixed by disabling animation from the start and only allowing
it after the centering has aleady happened.

As a result, I was able to cleanup few methods and places where
centering was needlessly called.