fix: swipe by pixels only on loop
Before, the distance to which the loop would shift to reposition the
slider was dependent on the selected index. But it's not that simple,
the selection can be different depending on whether the threshold
for item selection was crossed or not. But the tweak didn't take this
into account. So I decided to ignore the new selected index and instead
tweak only, which is simpler and more stable.