feat: improve infinte loop
The motivation for this is to simplify the looping mechanism by using
native api instead of faking fringe items in extensive amount. All this
to prepare for a move into a more general slider direction than just
image gallery.

Ideally, I'd use only the IntersectionObserver for the looping exclusively
but it is not practical. The problem is the selection with
IntersectionObserver is very much timing dependent, based on the animation
that at some point reveals fringe item and loops. Interference with user
swiping is unavoidable and handling conflicts between these would require
Web Animation API with its stopping capability. This is too complicated so
I resolved to keep control - use Inter Obs only for swiping, and the old
pointer-based mechanism for the rest of the looping scenarios. Doing this
removes the dependency on animation timing and instead bases the looping
solely on the pointer position, which is very much desired since it's
deterministic.

This commit also improves selection of items by removing the selection
bias, which was skewed in just 1 direction.

This commit also fixes a bug in the looping which didn't work correctly
when using arrows for looping. The loop did happen but was animated and
therefore ugly.