feat: add smoother multi-item navigation WIP
This commit introduces paging-like navigation which takes a desired
index for selection only as a guidance to choose a page, not as absolute
target. This should make it easier for user to orient himself in the
multitude of items and also to spot the end of the list.

Since the actual selection differs a little from the desired index
concept, upon which the gallery was based until now, it probably doesn't
make sense to track the selection anymore. For instance, given 4 items
in total, if a client would call selection with index 3, the gallery can
in the end select index 1 so that the last "page" is full. If tracked by
@Output, this would be confusing to the client, so it probably makes
sense to hide it and remove the outgoing triggers. Will be done in next
commit unless I find a reason to keep it.

This commit has the unexpected benefit of solving the fringe flicker as
described in https://trello.com/c/BS2TpM4w. It is also preparation for
the dots component.

The commit doesn't include tests because the Karma test suite has such a
bad DX I didn't want to waste my time with it.