fix: prevent looping being stuck in disabled state
Previously, when looping was disabled due to insufficient amount of
items, even if the items eventually arrived the looping would be stuck
in disabled state for ever. It worked so because the `loop` setting was
also holding current state, so the setting was easily overriden and the
original intent lost.

This commit holds those 2 pieces of info separate and as such makes it
possible to reactivate looping once the items arrive.

Because of the changes in ngOnChanges, I had to fix the way some tests
deal with ngOnChanges. Angular doesn't trigger this lifecycle method for
top level components automatically and each input that's handled in
ngOnChanges has to be included in a manual call. To make the job easier
I started using `setInput` which takes care of this automatically. Even
though it's a step forward, it's still miles from how easy it'd be to
test this case in Cypress component test, because I still have to be
aware of which inputs I need to call `setInput` for in my unit tests,
whereas in Cy component test I wouldn't have to deal with this low level
stuff.