fix: check that changes are defined
It can easily happen that items are undefined as a change when
ngOnChanges is called in GalleryComponent. This is a deficiency of the
Angular's types that this is not checked, and it slipped through my test
suite as well.

So I created a stricter type that will keep the possibility of changes
being undefined in check and reorganized the test so that the ngOnChanges
with undefined `items` change is checked as well.