build: cache node_modules in CI
I found that the `cache` I was using in setup-node job is not what I
though it is. I thought it caches the node_modules in CI so that they
can be then restored locally, but no no. It caches packages in a remote
global cache, which doesn't bring much benefit afaik.

Therefore, I use actions/cache now to have the local caching I
originally desired, and I don't use the setup-node cache until I find a
compelling reason for it.