test(gallery): improve testing config
First, I preselect Electron as the running browser, which skips the
manual browser selection step locally.

Then I set the config for command timeout to 1s instead of 4s, because
it makes the testing flow faster and there is no reason to wait that
long.

Lastly I set the waitForAnimations to false, but to me it doesn't seem
like it has an effect. It's purpose is to either wait or not for an
element to finish animating before running commands.