fix(gallery): re-center when items change in loop mode
Before, when items came in later, they were not re-centered and the
fringe item(s) was displayed. So I am centering it and turning off the
animation at the start so that it doesn't look awkward.

I tried to test this last aspect in Cypress but without avail, even when
turning off waitForAnimations. Given Cypress so far sucks for my use
case, since the hot reloading doesn't work, I might consider Playwright.